+++ src/TestTrackingDiagrams.Extensions.MassTransit/MassTransitTracker.cs
@@ -99,7 +99,8 @@ public class MassTransitTracker : ITrackingComponent
             label, body, uri, [],
             _options.ServiceName, _options.CallingServiceName,
             RequestResponseType.Request, traceId, requestResponseId, false,
-            MetaType: RequestResponseMetaType.Event)
+            MetaType: RequestResponseMetaType.Event,
+            DependencyCategory: "MessageQueue")
         {
             Phase = TestPhaseContext.Current
         });
@@ -109,7 +110,8 @@ public class MassTransitTracker : ITrackingComponent
             label, null, uri, [],
             _options.ServiceName, _options.CallingServiceName,
             RequestResponseType.Response, traceId, requestResponseId, false,
-            MetaType: RequestResponseMetaType.Event)
+            MetaType: RequestResponseMetaType.Event,
+            DependencyCategory: "MessageQueue")
         {
             Phase = TestPhaseContext.Current
         });
@@ -137,7 +139,8 @@ public class MassTransitTracker : ITrackingComponent
             label, body, uri, [],
             _options.CallingServiceName, _options.ServiceName,
             RequestResponseType.Request, traceId, requestResponseId, false,
-            MetaType: RequestResponseMetaType.Event)
+            MetaType: RequestResponseMetaType.Event,
+            DependencyCategory: "MessageQueue")
         {
             Phase = TestPhaseContext.Current
         });
@@ -147,7 +150,8 @@ public class MassTransitTracker : ITrackingComponent
             label, null, uri, [],
             _options.CallingServiceName, _options.ServiceName,
             RequestResponseType.Response, traceId, requestResponseId, false,
-            MetaType: RequestResponseMetaType.Event)
+            MetaType: RequestResponseMetaType.Event,
+            DependencyCategory: "MessageQueue")
         {
             Phase = TestPhaseContext.Current
         });
@@ -173,7 +177,8 @@ public class MassTransitTracker : ITrackingComponent
             label, exception.Message, uri, [],
             svc, caller,
             RequestResponseType.Request, traceId, requestResponseId, false,
-            MetaType: RequestResponseMetaType.Event)
+            MetaType: RequestResponseMetaType.Event,
+            DependencyCategory: "MessageQueue")
         {
             Phase = TestPhaseContext.Current
         });
@@ -184,7 +189,8 @@ public class MassTransitTracker : ITrackingComponent
             svc, caller,
             RequestResponseType.Response, traceId, requestResponseId, false,
             StatusCode: "Fault",
-            MetaType: RequestResponseMetaType.Event)
+            MetaType: RequestResponseMetaType.Event,
+            DependencyCategory: "MessageQueue")
         {
             Phase = TestPhaseContext.Current
         });